Add forSuggestions attribute to CommandContextBuilder
Currently parsing and execution/suggestions are performed separately.
This separation of concerns is good, but sometimes can lead to duplicate
work. For example, it is necessary to perform a duplicate permissions
checking call when creating the suggestions for an argument node.
Introducing this change as a boolean field instead of modifying the
parsing logic leads to better maintainability.